requirements.txt
python-can==4.3.0
streamlit==1.29.0
pandas==2.0.3
plotly==5.17.0
python-dotenv==1.0.0